% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphstreamjars-package.R
\docType{package}
\name{graphstreamjars}
\alias{graphstreamjars}
\title{Java Archive Wrapper Supporting the 'graphstream' Package}
\description{
Contains all supporting JARs for working with the GraphStream Java APIs.
(\url{http://graphstream-project.org/})Version number matches version number of included
'graphstreamjars' libraries.
}
\author{
Bob Rudis (bob@rud.is)
}
